Use the autodiscover_modules to load modules with cron.py
Before it was trying to import .cron from all apps from INSTALLED_APPS,
but this missed the new AppConfig apps, which are in INSTALLED_APPS as
objects and not modules. Django uses autodiscover_modules internally
for models.py (for example), so it should work on all expected situations.

I did left the previous PROJECT_MODULE.__name__ as I dont fully
understand why it is there.