cron load module discovery fix
The PROJECT_MODULE calculation in settings is causing KeyError for me. autodiscover_modules is probably a good solution but not available on Django 1.6.

What I did was to perform the old cron module discovery if autodiscover_modules is not available and If a KeyError appens, PROJECT_MODULE is set to None